test(startup): split over-limit ApplicationGlobals tests; close #202 review findings
Remediation cycle 1 for the feature-review of #202.

- Finding 1 (blocking): ApplicationGlobalsTests.cs exceeded the 500-line
  file-size limit (687 lines). Mechanically extract the four startup-timing
  wiring tests and their helpers into a new
  ApplicationGlobalsStartupTimingTests.cs. Both files are now under 500
  lines (483 and 299). The split preserves [DoNotParallelize] and the
  Settings.Default.StartupTimingEnabled save/restore; no assertion or test
  intent changed.
- Finding 2 (non-blocking): emit merged Cobertura to artifacts/csharp/coverage.xml
  to satisfy the feature-review-workflow artifact contract.

Also records the cycle-1 audit artifacts (policy-audit, code-review,
feature-audit, remediation-inputs, remediation-plan).

Full C# toolchain green in a single pass; 4194/4194 tests pass; new-code
coverage 100%; repo-wide first-party coverage unchanged (no regression).
