+++ docs/features/active/2026-06-15-debug-startup-timing-instrumentation-202/evidence/baseline/phase0-instructions-read.2026-06-15T13-29.md
@@ -1,0 +1,22 @@
+# Phase 0 — Policy Read Evidence (Issue #202, Remediation Cycle 2026-06-15T13-29)
+
+Timestamp: 2026-06-15T13-29
+
+Policy Order: per `policy-compliance-order` — (1) CLAUDE.md, (2) `.claude/rules/general-code-change.md`, (3) `.claude/rules/general-unit-test.md`, (4) language-specific `.claude/rules/csharp.md`.
+
+Files read (in mandatory order):
+
+1. `CLAUDE.md`
+2. `.claude/rules/general-code-change.md`
+3. `.claude/rules/general-unit-test.md`
+4. `.claude/rules/csharp.md`
+
+Summary: All four policy files were read in the mandatory order before any execution.
+This remediation cycle performs a pure mechanical test-file split (no production-code
+change, no assertion or test-intent change) plus a non-blocking coverage-artifact copy.
+Applicable constraints: 500-line file-size limit (General Code Change Policy §4),
+CSharpier-only formatting, .NET analyzers + nullable/TreatWarningsAsErrors gates,
+MSTest + Moq + FluentAssertions, vstest with `/EnableCodeCoverage`, banned APIs
+(`DateTime.Now/UtcNow`, `Thread.Sleep`, `Task.Delay`, `Random.Shared`), no temp files
+in tests, preserve `[DoNotParallelize]` and `Settings.Default.StartupTimingEnabled`
+save/restore.
