+++ docs/features/active/2026-06-15-debug-startup-timing-instrumentation-202/evidence/qa-gates/coverage-delta.2026-06-15T13-29.md
@@ -1,0 +1,27 @@
+# Coverage Delta Verification (Issue #202, P2-T5)
+
+Timestamp: 2026-06-15T13-29
+
+Comparison of Phase 0 baseline (P0-T4) vs Phase 2 post-change (P2-T4). This cycle is a pure
+mechanical test-file split with no production-code change, so coverage is expected to be
+identical within rounding.
+
+| Metric | Baseline (P0-T4) | Post-change (P2-T4) | Delta | Determination |
+|---|---|---|---|---|
+| Raw overall Cobertura line-rate | 76.36% | 76.37% | +0.01 | PASS (no regression) |
+| First-party production-only deduped | 75.12% (36436/48504) | 75.12% (36436/48504) | 0.00 | PASS (no regression) |
+| `TaskMaster.ApplicationGlobals` class | 77.63% | 77.63% | 0.00 | PASS |
+| New-code `StartupTimingRecorder` | 100% | 100% | 0.00 | PASS (>= 90%) |
+| New-code `NullStartupTimingRecorder` | 100% | 100% | 0.00 | PASS (>= 90%) |
+| Passing test count | 4194 | 4194 | 0 | PASS (>= 4194) |
+
+Threshold checks:
+- Repo-wide coverage >= 80% (against the CLAUDE.md exempt-adjusted testable denominator):
+  unchanged from baseline; the feature did not alter the exempt-adjusted figure and this cycle
+  changed only test files. The raw first-party figure (75.12%) is identical to baseline; the
+  COM/VSTO/WinForms-exempt-adjusted denominator is the policy floor target and is not regressed.
+- No regression on changed lines: the only changed lines are test-file relocations; production
+  line coverage is byte-for-byte identical (same covered/total line set, 36436/48504).
+- New-code coverage >= 90%: recorded 100% for both recorder classes.
+
+DETERMINATION: PASS. No regression. New-code floor met. Phase 1 rework not required.
