+++ docs/features/active/2026-06-15-debug-startup-timing-instrumentation-202/evidence/qa-gates/qa-format.2026-06-15T13-29.md
@@ -1,0 +1,18 @@
+# QA Gate — Format (CSharpier) (Issue #202, P2-T1)
+
+Timestamp: 2026-06-15T13-29
+
+Command: `csharpier format .`
+
+EXIT_CODE: 0
+
+Output Summary:
+
+- `Formatted 1058 files in 747ms.` EXIT_CODE 0.
+- `git status --short` confirms only the two intended C# files are modified
+  (`TaskMaster.Test/AppGlobals/ApplicationGlobalsTests.cs`,
+  `TaskMaster.Test/TaskMaster.Test.csproj`) plus the new untracked
+  `TaskMaster.Test/AppGlobals/ApplicationGlobalsStartupTimingTests.cs`. CSharpier did not
+  rewrite any other tracked source file.
+- The two split files were already CSharpier-clean from the targeted format run; the full-repo
+  pass produced no further change to them. No loop restart required.
