+++ docs/features/active/2026-06-15-debug-startup-timing-instrumentation-202/evidence/qa-gates/verification-summary.2026-06-15T13-29.md
@@ -1,0 +1,61 @@
+# Verification Summary (Issue #202, Phase 4)
+
+Timestamp: 2026-06-15T13-29
+
+## P4-T1 — File-size finding closed
+
+Post-split line counts (post-CSharpier, from P1-T7):
+- `TaskMaster.Test/AppGlobals/ApplicationGlobalsTests.cs`: 483 lines (< 500).
+- `TaskMaster.Test/AppGlobals/ApplicationGlobalsStartupTimingTests.cs`: 299 lines (< 500).
+
+Both files are strictly under the 500-line limit. Finding 1 (BLOCKING) is closed.
+
+## P4-T2 — No test loss
+
+- Phase 2 post-change passing count (P2-T4): 4194 / 4194 (>= 4194 required). Failed: 0.
+- The four `[DoNotParallelize]` startup-timing wiring tests all appear in the run results under
+  the new `ApplicationGlobalsStartupTimingTests` class and all passed:
+  `LoadAsync_WhenTimingDisabled_RecordsNothingAndEmitsNoTable`,
+  `LoadAsync_WhenTimingEnabled_RecordsAllPhasesInStartupOrderWithLoadBasicFirst`,
+  `LoadAsync_WhenTimingEnabled_EmitsExactlyOneTableWithPhaseNamesAndTotal`,
+  `LoadAsync_PreservesPhaseOrderingAndYieldCount_WhenTimingOnVersusOff`.
+- The new file retains all four `[DoNotParallelize]` markers (4) and the
+  `Settings.Default.StartupTimingEnabled` save/restore lifecycle; the original retains 0.
+
+## P4-T3 — Assertion / intent parity
+
+- `git diff` of the original `ApplicationGlobalsTests.cs` is pure deletions: 204 deletions, 0
+  insertions. No assertion, attribute, or comment in the retained tests was altered.
+- A line-by-line diff of the moved region (git HEAD original lines 365-498) against the moved
+  test methods in the new file shows the bodies are byte-identical; the only difference is one
+  trailing closing brace captured by the extraction window. The only cross-file differences are
+  file relocation, the new class name, and the necessarily-duplicated helpers
+  (`SetEnginesMock`, `AttachMemoryAppender`, `DetachMemoryAppender`, `CreateOutlookApplicationStub`,
+  and the nested `TestableApplicationGlobals` with its `TimingRecorder` seam / `LoadBasicMethod`
+  override / phase overrides / `YieldCount`). No assertion text changed.
+
+## P4-T4 — Coverage floors and AC validity
+
+- Repo-wide / first-party production-only deduped coverage: 75.12% (36436/48504) post-change,
+  identical to baseline. No regression on changed lines (changed lines are test-file
+  relocations only).
+- New-code coverage: `StartupTimingRecorder` 100%, `NullStartupTimingRecorder` 100% (>= 90%).
+- All five acceptance criteria in `spec.md` (lines 91-95) and `user-story.md` (lines 50-54)
+  remain `[x]` and PASS; the mechanical split does not affect AC delivery. No checkbox change
+  required.
+
+## P4-T5 — Findings addressed
+
+- Finding 1 (BLOCKING — test-file > 500 lines): CLOSED. Both files < 500 lines after the split.
+- Finding 2 (NON-BLOCKING — missing `artifacts/csharp/coverage.xml`): CLOSED. The merged
+  Cobertura was copied to `artifacts/csharp/coverage.xml` (P3) and parses to the P2-T4 figures.
+- Blocking-finding count after remediation: 0.
+
+## Final Toolchain Result (single clean pass)
+
+1. Format (`csharpier format .`): EXIT_CODE 0.
+2. Analyze (`msbuild ... /p:EnableNETAnalyzers=true /p:EnforceCodeStyleInBuild=true`): EXIT_CODE 0.
+3. Type-check/nullable (`msbuild ... /p:Nullable=enable /p:TreatWarningsAsErrors=true`): EXIT_CODE 0.
+4. Test+coverage (`vstest.console.exe ... /EnableCodeCoverage /InIsolation`): EXIT_CODE 0, 4194/4194.
+
+No step changed files or failed; the loop completed in a single pass.
