+++ docs/features/active/2026-06-15-debug-startup-timing-instrumentation-202/evidence/remediation-baseline/baseline-linecounts.2026-06-15T13-29.md
@@ -1,0 +1,22 @@
+# Remediation Baseline — Line Counts and Target-File Absence (Issue #202)
+
+Timestamp: 2026-06-15T13-29
+
+## [P0-T2] Baseline line count of affected test file
+
+Command: `awk 'END{print NR}' TaskMaster.Test/AppGlobals/ApplicationGlobalsTests.cs`
+
+EXIT_CODE: 0
+
+Output Summary: 687 lines. This exceeds the 500-line file-size limit (Finding 1, BLOCKING),
+matching the expected baseline of 687.
+
+## [P0-T3] New target file absence confirmation
+
+Command: `test -f TaskMaster.Test/AppGlobals/ApplicationGlobalsStartupTimingTests.cs; echo $?`
+
+EXIT_CODE: 1 (file absent)
+
+Output Summary: The new target file
+`TaskMaster.Test/AppGlobals/ApplicationGlobalsStartupTimingTests.cs` does not exist before
+the split (`test -f` returned exit code 1). The split will create it.
